+++ src/Scripts/FunctionMappings/SurfaceLogic.lua
@@ -5,6 +5,11 @@ function mod.RoomEntranceE_Intro(currentRun, currentRoom)
 	local roomIntroSequenceDuration = currentRoom.IntroSequenceDuration or game.RoomData.BaseRoom.IntroSequenceDuration or
 			0.0
 
+	-- Prevent Hephaestus blast VFX
+	game.RemoveReadiedMassiveAttackPresentation("HephaestusWeaponBoon")
+	game.RemoveReadiedMassiveAttackPresentation("HephaestusSpecialBoon")
+	game.RemoveReadiedMassiveAttackPresentation("HephaestusSprintBoon")
+
 	FadeOut({ Color = game.Color.Black, Duration = 0 })
 	AdjustFullscreenBloom({ Name = "LightningStrike", Duration = 0 })
 	AdjustFullscreenBloom({ Name = "WrathPhase2", Duration = 0.1, Delay = 0 })
@@ -228,6 +233,12 @@ function mod.RoomEntranceSurface(currentRun, currentRoom)
 	game.HideCombatUI("Surface")
 	game.ZeroMouseTether("RoomEntrance")
 	AdjustZoom({ Fraction = game.CurrentRun.CurrentRoom.IntroZoomFraction or 0.7, Duration = 0.0 })
+
+	-- Prevent Hephaestus blast VFX
+	game.RemoveReadiedMassiveAttackPresentation("HephaestusWeaponBoon")
+	game.RemoveReadiedMassiveAttackPresentation("HephaestusSpecialBoon")
+	game.RemoveReadiedMassiveAttackPresentation("HephaestusSprintBoon")
+
 	FadeIn({ Duration = 5.5 })
 	PanCamera({ Id = currentRoom.CameraEndPoint, Duration = currentRoom.IntroSequenceDuration or 4, EaseIn = 0.0, EaseOut = 1.5 })
 	game.FullScreenFadeInAnimation("RoomTransitionOutBlack")